Change endorsed logo widget option for RSS feed.
Only way to guarantee that a site with no posts will not produce a 404 error when accessing the RSS feed link is to prevent it from appearing in the header. If an additional option to add an RSS icon in the SM footer widget is desired, I’ll add it back.